EDMT-75 refactor : 전체 학생 데이터 넣는 로직 이름 수정
